requirements.txt
streamlit
openai
llama-index
boto3
paramiko
gitpython
pandas
